requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.4
djangorestframework
django-toolbelt==0.0.1
gunicorn==19.4.5
psycopg2==2.6.1
python-decouple==3.0
static3==0.6.1
wheel==0.29.0
Pillow
